v0.2.1
Front page completed
