--- src/modules/PageOne/Header.js
@@ -1,22 +1,0 @@
-import elFactory from "../../functions/elementFactory"
-import { card } from "../Cube"
-
-export const header = () => {
-    const header = elFactory("div", {class: "header"}, 
-        elFactory("div", {class: "header-square"}, 
-            elFactory("span", {}, "JAYDEN"),
-            elFactory("br", {}),
-            elFactory("span", {}, "REYNOLDS"),
-        )
-    )
-
-    // header.addEventListener("click", () => {
-    //     const cube = document.querySelector(".cube")
-    //     cube.classList.remove("show-front")
-    //     cube.classList.add("show-right")
-    // })
-
-    return (
-        header
-    )
-}
