--- src/modules/Pages/PageFour.js
@@ -1,7 +1,0 @@
-import elFactory from "../../functions/elementFactory"
-
-export const body = () => {
-    const page = elFactory("div", {class: "page-four cube__face cube__face--back"}, "Test")
-
-    return page
-}
