--- src/modules/Pages/PageThree.js
@@ -1,7 +1,0 @@
-import elFactory from "../../functions/elementFactory"
-
-export const body = () => {
-    const page = elFactory("div", {class: "page-three cube__face cube__face--left "}, "Test")
-
-    return page
-}
